Refactor create and edit IPO
