+++ src/modules/InvitationForPunchOut/views/CreateAndEditIPO/Participants/__tests__/Participants.test.jsx
@@ -3,24 +3,6 @@ import { render, waitFor } from '@testing-library/react';
 import Participants from '../Participants';
 import React from 'react';
 
-const mockRoles = [
-    {
-        code: 'Code a',
-        description: 'Desc',
-        email: 'codea@test.com',
-        informationalEmail: null,
-        usePersonalEmail: false,
-        persons: [ 
-            {
-                azureOid: '00-11-22',
-                firstName: 'Elisabeth',
-                lastName: 'Bratli',
-                email: 'elisabeth@email.com'
-            }
-        ]
-    }
-];
-
 const participants = [
     {
         organization: 'Contractor',
@@ -35,16 +17,6 @@ const participants = [
         role: null
     }
 ];
-
-jest.mock('../../../../context/InvitationForPunchOutContext',() => ({
-    useInvitationForPunchOutContext: () => {
-        return {
-            apiClient: {
-                getFunctionalRolesAsync: () => Promise.resolve(mockRoles)
-            }
-        };
-    }
-}));
 
 describe('Module: <Participants />', () => {
