+++ src/modules/InvitationForPunchOut/views/CreateAndEditIPO/__tests__/CreateIPO.test.tsx
@@ -1,11 +1,28 @@
 import { fireEvent, render, waitFor, within } from '@testing-library/react';
-
-import CreateAndEditIPO from '../CreateAndEditIPO';
 import React from 'react';
 import { act } from 'react-dom/test-utils';
 import { configure } from '@testing-library/dom';
+import CreateIPO from '../CreateIPO';
 
 configure({ testIdAttribute: 'id' }); // makes id attibute data-testid for subsequent tests
+
+const mockRoles = [
+    {
+        code: 'Code a',
+        description: 'Desc',
+        email: 'codea@test.com',
+        informationalEmail: null,
+        usePersonalEmail: false,
+        persons: [
+            {
+                azureOid: '00-11-22',
+                firstName: 'Elisabeth',
+                lastName: 'Bratli',
+                email: 'elisabeth@email.com'
+            }
+        ]
+    }
+];
 
 jest.mock('react-router-dom', () => ({
     useParams: (): { projectId: any, commPkgNo: any } => ({
@@ -24,6 +41,7 @@ jest.mock('../../../context/InvitationForPunchOutContext', () => ({
     useInvitationForPunchOutContext: (): any => {
         return {
             apiClient: {
+                getFunctionalRolesAsync: (): any => Promise.resolve(mockRoles),
                 getAllProjectsForUserAsync: (_: any): any => {
                     return Promise.resolve([
                         {
@@ -50,62 +68,73 @@ jest.mock('@procosys/core/DirtyContext', () => ({
     }
 }));
 
-describe('<CreateAndEditIPO />', () => {
+describe('<CreateIPO />', () => {
     afterEach(() => {
         jest.clearAllMocks();
     });
 
     it('Should display type types when Type of punch round select is clicked', async () => {
-        const result = render(<CreateAndEditIPO />);
         await act(async () => {
+            const result = render(<CreateIPO />);
             const selectContainer = result.getByTestId('po-type-select');
-
             const utils = within(selectContainer);
             const button = utils.getByRole('listbox');
             fireEvent.click(button);
+            await waitFor(() => expect(result.getByText('DP (Discipline Punch)')).toBeInTheDocument());
+            await waitFor(() => expect(result.getByText('MDP (Multi Discipline Punch)')).toBeInTheDocument());
         });
-
-        expect(result.getByText('DP (Discipline Punch)')).toBeInTheDocument();
-        expect(result.getByText('MDP (Multi Discipline Punch)')).toBeInTheDocument();
     });
 
     it('Should render Next button enabled', async () => {
-        const { getByText } = render(<CreateAndEditIPO />);
-        await waitFor(() => expect(getByText('Next').closest('button')).toHaveProperty('disabled', true));
+        await act(async () => {
+            const { getByText } = render(<CreateIPO />);
+            expect(getByText('Next').closest('button')).toHaveProperty('disabled', true);
+        });
     });
 
     it('Should render Previous button enabled', async () => {
-        const { getByText } = render(<CreateAndEditIPO />);
-        await waitFor(() => expect(getByText('Previous').closest('button')).toHaveProperty('disabled', true));
+        await act(async () => {
+            const { getByText } = render(<CreateIPO />);
+            await waitFor(() => expect(getByText('Previous').closest('button')).toHaveProperty('disabled', true));
+        });
     });
 
     it('Should update end time when entering start time > end time', async () => {
-        const { getByTestId } = render(<CreateAndEditIPO />);
-        const startTime = getByTestId('startTime');
-        const endTime = getByTestId('endDate');
-        fireEvent.change(startTime, { target: { value: '08:00' } });
-        fireEvent.change(endTime, { target: { value: '09:00' } });
-        fireEvent.change(startTime, { target: { value: '22:23' } });
-        await waitFor(() => expect(endTime).toHaveValue('23:23'));
+        await act(async () => {
+            const { getByTestId } = render(<CreateIPO />);
+            const startTime = getByTestId('startTime');
+            const endTime = getByTestId('endDate');
+            fireEvent.change(startTime, { target: { value: '08:00' } });
+            fireEvent.change(endTime, { target: { value: '09:00' } });
+            fireEvent.change(startTime, { target: { value: '22:23' } });
+            await waitFor(() => expect(endTime).toHaveValue('23:23'));
+        });
     });
 
     it('Should not update end time when entering start time < end time', async () => {
-        const { getByTestId } = render(<CreateAndEditIPO />);
-        const startTime = getByTestId('startTime');
-        const endTime = getByTestId('endDate');
-        fireEvent.change(startTime, { target: { value: '08:00' } });
-        fireEvent.change(endTime, { target: { value: '23:23' } });
-        fireEvent.change(startTime, { target: { value: '10:23' } });
-        await waitFor(() => expect(endTime).toHaveValue('23:23'));
+        await act(async () => {
+            const { getByTestId } = render(<CreateIPO />);
+            const startTime = getByTestId('startTime');
+            const endTime = getByTestId('endDate');
+            fireEvent.change(startTime, { target: { value: '08:00' } });
+            fireEvent.change(endTime, { target: { value: '23:23' } });
+            fireEvent.change(startTime, { target: { value: '10:23' } });
+            waitFor(() => expect(endTime).toHaveValue('23:23'));
+        });
     });
 
+
     it('Should not update time when invalid (empty)', async () => {
-        const { getByTestId } = render(<CreateAndEditIPO />);
-        const startTime = getByTestId('startTime');
-        fireEvent.change(startTime, { target: { value: '13:30' } });
-        fireEvent.change(startTime, { target: { value: '' } });
-        await waitFor(() => expect(startTime).toHaveValue('13:30'));
+        await act(async () => {
+            const { getByTestId } = render(<CreateIPO />);
+            const startTime = getByTestId('startTime');
+            fireEvent.change(startTime, { target: { value: '13:30' } });
+            fireEvent.change(startTime, { target: { value: '' } });
+            await waitFor(() => expect(startTime).toHaveValue('13:30'));
+
+        });
     });
+
 });
